backend/FruitClassifier/requirements.txt
Pillow==8.4.0
tensorflow
typing-extensions==4.1.1
keras
fastapi
uvicorn
python-multipart==0.0.5